Consolidated Entities into their own respective header and source filters. This cleans up the structure a bit more.
