+++ Headers/EntityHeaders/SpotLight.h
