+++ Headers/EntityManager.h
@@ -1,14 +1,14 @@
 #pragma once
 #include "stdafx.h"
-#include "PointLight.h"
-#include "DirectionalLight.h"
-#include "SpotLight.h"
+#include "EntityHeaders/PointLight.h"
+#include "EntityHeaders/DirectionalLight.h"
+#include "EntityHeaders/SpotLight.h"
 #include "Scene_Loader.h"
 #include "BoidEngine.h"
 #include "EmitterEngine.h"
-#include "InteractableEntity.h"
-#include "Entity.h"
-#include "Camera.h"
+#include "EntityHeaders/InteractableEntity.h"
+#include "EntityHeaders/Entity.h"
+#include "EntityHeaders/Camera.h"
 #include "EntityComponentHeaders/EntityComponent.h"
 #include "EntityComponentHeaders/CameraComponent.h"
 #include "EntityComponentHeaders/RenderComponent.h"
