+++ Renderer.vcxproj
@@ -21,33 +21,32 @@
   <ItemGroup>
     <ClInclude Include="Headers\Anim_Track.h" />
     <ClInclude Include="Headers\BoidEngine.h" />
-    <ClInclude Include="Headers\Camera.h" />
     <ClInclude Include="Headers\CommandHandler.h" />
-    <ClInclude Include="Headers\DirectionalLight.h" />
     <ClInclude Include="Headers\Emitter.h" />
     <ClInclude Include="Headers\EmitterEngine.h" />
-    <ClInclude Include="Headers\Entity.h" />
     <ClInclude Include="Headers\EntityComponentHeaders\CameraComponent.h" />
     <ClInclude Include="Headers\EntityComponentHeaders\EntityComponent.h" />
     <ClInclude Include="Headers\EntityComponentHeaders\LightingComponent.h" />
     <ClInclude Include="Headers\EntityComponentHeaders\RenderComponent.h" />
+    <ClInclude Include="Headers\EntityHeaders\Camera.h" />
+    <ClInclude Include="Headers\EntityHeaders\DirectionalLight.h" />
+    <ClInclude Include="Headers\EntityHeaders\Entity.h" />
+    <ClInclude Include="Headers\EntityHeaders\InteractableEntity.h" />
+    <ClInclude Include="Headers\EntityHeaders\PlayerEntity.h" />
+    <ClInclude Include="Headers\EntityHeaders\PointLight.h" />
+    <ClInclude Include="Headers\EntityHeaders\SpotLight.h" />
+    <ClInclude Include="Headers\EntityHeaders\StaticEntity.h" />
     <ClInclude Include="Headers\EntityManager.h" />
     <ClInclude Include="Headers\EnvSpec.h" />
-    <ClInclude Include="Headers\FreeImage.h" />
     <ClInclude Include="Headers\GameTime.h" />
     <ClInclude Include="Headers\GameManager.h" />
     <ClInclude Include="Headers\ImageReader.h" />
     <ClInclude Include="Headers\InputHandler.h" />
-    <ClInclude Include="Headers\InteractableEntity.h" />
-    <ClInclude Include="Headers\PointLight.h" />
     <ClInclude Include="Headers\Mesh.h" />
     <ClInclude Include="Headers\MeshManager.h" />
     <ClInclude Include="Headers\Scene_Loader.h" />
-    <ClInclude Include="Headers\PlayerEntity.h" />
     <ClInclude Include="Headers\Shader.h" />
     <ClInclude Include="Headers\ShaderManager.h" />
-    <ClInclude Include="Headers\SpotLight.h" />
-    <ClInclude Include="Headers\StaticEntity.h" />
     <ClInclude Include="Headers\stb_image.h" />
     <ClInclude Include="Headers\stdafx.h" />
     <ClInclude Include="Headers\Texture.h" />
@@ -58,12 +57,17 @@
   <ItemGroup>
     <ClCompile Include="Source\Anim_Track.cpp" />
     <ClCompile Include="Source\BoidEngine.cpp" />
-    <ClCompile Include="Source\Camera.cpp" />
     <ClCompile Include="Source\CommandHandler.cpp" />
-    <ClCompile Include="Source\DirectionalLight.cpp" />
     <ClCompile Include="Source\Emitter.cpp" />
     <ClCompile Include="Source\EmitterEngine.cpp" />
-    <ClCompile Include="Source\Entity.cpp" />
+    <ClCompile Include="Source\Entities\Camera.cpp" />
+    <ClCompile Include="Source\Entities\DirectionalLight.cpp" />
+    <ClCompile Include="Source\Entities\Entity.cpp" />
+    <ClCompile Include="Source\Entities\InteractableEntity.cpp" />
+    <ClCompile Include="Source\Entities\PlayerEntity.cpp" />
+    <ClCompile Include="Source\Entities\PointLight.cpp" />
+    <ClCompile Include="Source\Entities\SpotLight.cpp" />
+    <ClCompile Include="Source\Entities\StaticEntity.cpp" />
     <ClCompile Include="Source\EntityComponents\CameraComponent.cpp" />
     <ClCompile Include="Source\EntityComponents\EntityComponent.cpp" />
     <ClCompile Include="Source\EntityComponents\LightingComponent.cpp" />
@@ -73,17 +77,12 @@
     <ClCompile Include="Source\GameManager.cpp" />
     <ClCompile Include="Source\ImageReader.cpp" />
     <ClCompile Include="Source\InputHandler.cpp" />
-    <ClCompile Include="Source\InteractableEntity.cpp" />
-    <ClCompile Include="Source\PointLight.cpp" />
     <ClCompile Include="Source\main.cpp" />
     <ClCompile Include="Source\Mesh.cpp" />
     <ClCompile Include="Source\MeshManager.cpp" />
     <ClCompile Include="Source\Scene_Loader.cpp" />
-    <ClCompile Include="Source\PlayerEntity.cpp" />
     <ClCompile Include="Source\Shader.cpp" />
     <ClCompile Include="Source\ShaderManager.cpp" />
-    <ClCompile Include="Source\SpotLight.cpp" />
-    <ClCompile Include="Source\StaticEntity.cpp" />
     <ClCompile Include="Source\Texture.cpp" />
     <ClCompile Include="Source\TextureManager.cpp" />
     <ClCompile Include="Source\UserInterface.cpp" />
