+++ Source/Entities/Entity.cpp
@@ -1,4 +1,4 @@
-#include "Entity.h"
+#include "EntityHeaders/Entity.h"
 
 // Default Constructor
 Entity::Entity(int iID, vec3 vPosition)
